captions and degree row
