+++ app/components/caption_text.html.erb
