--- app/components/caption_text/view.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module CaptionText
-  class View < ViewComponent::Base
-    attr_reader :text
-    def initialize(text:)
-      @text = text
-      super
-    end
-  end
-end
