+++ app/components/degree_row_content.html.erb
