+++ app/views/publish/courses/_description_content.html.erb
@@ -126,7 +126,7 @@
       summary_list,
       :course,
       "Degree",
-      (render DegreeRowContentComponent::View.new(course:, errors: @errors)),
+      (render DegreeRowContent.new(course:, errors: @errors)),
       %w[degree_grade degree_subject_requirements],
       truncate_value: false,
       action_path: course.degree_section_complete? ? degrees_start_publish_provider_recruitment_cycle_course_path(@provider.provider_code, @provider.recruitment_cycle_year, course.course_code) : nil,
