+++ app/views/publish/courses/accredited_body/new.html.erb
@@ -9,7 +9,7 @@
 <fieldset class="govuk-fieldset">
   <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
     <h1 class="govuk-fieldset__heading">
-      <%= render CaptionText::View.new(text: t("course.add_course")) %>
+      <%= render CaptionText.new(text: t("course.add_course")) %>
       Who is the accredited body?
     </h1>
   </legend>
