+++ app/views/publish/courses/apprenticeship/edit.html.erb
@@ -8,7 +8,7 @@
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-full">
-    <%= render CaptionText::View.new(text: course.name_and_code) %>
+    <%= render CaptionText.new(text: course.name_and_code) %>
     <%= form_with model: @course_funding_form,
                   url: apprenticeship_publish_provider_recruitment_cycle_course_path(@course.provider_code, @course.recruitment_cycle_year, @course.course_code),
                   method: :put do |form| %>
