+++ app/views/publish/courses/degrees/start/edit.html.erb
@@ -6,7 +6,7 @@
   <%= govuk_back_link_to(publish_provider_recruitment_cycle_course_path) %>
 <% end %>
 
-<%= render CaptionText::View.new(text: @course.name_and_code) %>
+<%= render CaptionText.new(text: @course.name_and_code) %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
