+++ app/views/publish/courses/engineers_teach_physics/_form_fields.html.erb
@@ -3,7 +3,7 @@
     <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
       <h1 class="govuk-fieldset__heading">
       <% if course.course_code %>
-          <%= render CaptionText::View.new(text: course.name_and_code) %>
+          <%= render CaptionText.new(text: course.name_and_code) %>
         <% end %>
         Engineers Teach Physics
       </h1>
