+++ app/views/publish/courses/sites/edit.html.erb
@@ -20,7 +20,7 @@
 <% end %>
 
 <h1 class="govuk-heading-l">
-  <%= render CaptionText::View.new(text: course.name_and_code) %>
+  <%= render CaptionText.new(text: course.name_and_code) %>
   Select the locations for this course
 </h1>
