+++ app/views/publish/courses/sites/new.html.erb
@@ -14,7 +14,7 @@
       <fieldset class="govuk-fieldset">
         <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
           <h1 class="govuk-fieldset__heading" data-qa="page-heading">
-            <%= render CaptionText::View.new(text: t("course.add_course")) %>
+            <%= render CaptionText.new(text: t("course.add_course")) %>
             Select the locations for this course
           </h1>
         </legend>
