+++ app/views/publish/courses/skilled_worker_visa_sponsorship/new.html.erb
@@ -8,7 +8,7 @@
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= render CaptionText::View.new(text: t("course.add_course")) %>
+    <%= render CaptionText.new(text: t("course.add_course")) %>
     <%= form_with(
           url:continue_publish_provider_recruitment_cycle_courses_skilled_worker_visa_sponsorship_path(
             @provider.provider_code,
