+++ app/views/publish/courses/start_date/new.html.erb
@@ -7,7 +7,7 @@
 <%= render "publish/shared/errors" %>
 
 <h1 class="govuk-heading-l">
-  <%= render CaptionText::View.new(text: t("course.add_course")) %>
+  <%= render CaptionText.new(text: t("course.add_course")) %>
   When does the course start?
 </h1>
