+++ app/views/publish/courses/student_visa_sponsorship/edit.html.erb
@@ -26,7 +26,7 @@
       <%= f.govuk_error_summary %>
 
       <h1 class="govuk-heading-l">
-        <%= render CaptionText::View.new(text: course.name_and_code) %>
+        <%= render CaptionText.new(text: course.name_and_code) %>
         <%= t("page_titles.student_visas.edit") %>
       </h1>
