+++ app/views/publish/courses/study_mode/edit.html.erb
@@ -25,7 +25,7 @@
 
       <%= f.govuk_error_summary %>
 
-      <%= render CaptionText::View.new(text: course.name_and_code) %>
+      <%= render CaptionText.new(text: course.name_and_code) %>
 
       <%= f.govuk_radio_buttons_fieldset(:study_mode, legend: { text: "Full time or part time?", tag: "h1", size: "l" }) do %>
         <% course.edit_course_options["study_modes"].each_with_index do |study_mode, index| %>
