+++ app/views/support/allocations/index.html.erb
@@ -5,5 +5,6 @@
 <%= render "support/menu" %>
 
 <%= render PaginatedFilter::View.new(filter_params: @filter_params, collection: @allocations) do %>
-  <%= render AllocationRecordCard::AllocationRecordCard.with_collection(@allocations, recruitment_cycle_year: params[:recruitment_cycle_year]) %>
+  <%= render AllocationRecordCard.with_collection(@allocations, recruitment_cycle_year: params[:recruitment_cycle_year]) %>
+
 <% end %>
